refactor: rename contribute-card to contribution-card and update import paths for consistency
