+++ web/src/components/search.tsx
@@ -5,10 +5,10 @@ import { ProjectEntity } from "@dzcode.io/models/dist/project";
 import { ContributorEntity } from "@dzcode.io/models/dist/contributor";
 import { ContributionEntity } from "@dzcode.io/models/dist/contribution";
 import { RepositoryEntity } from "@dzcode.io/models/dist/repository";
-import ContributionCard from "./contribute-card";
-import ContributorCard from "./contributor-card";
-import ProjectCard from "./project-card";
 import { useSearchModal } from "src/utils/search-modal";
+import { ContributionCard } from "./contribution-card";
+import { ContributorCard } from "./contributor-card";
+import { ProjectCard } from "./project-card";
 
 export function Search(): JSX.Element {
   const { localize } = useLocale();
